+++ src/main/java/com/almostreliable/summoningrituals/client/package-info.java
@@ -1,4 +1,0 @@
-@NonnullDefault
-package com.almostreliable.summoningrituals.client;
-
-import org.lwjgl.system.NonnullDefault;