+++ src/main/java/com/almostreliable/summoningrituals/client/tooltip/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.summoningrituals.client.tooltip;
+
+import org.lwjgl.system.NonnullDefault;