+++ src/main/java/com/almostreliable/summoningrituals/compat/viewer/jei/widget/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.summoningrituals.compat.viewer.jei.widget;
+
+import org.lwjgl.system.NonnullDefault;